requirements.txt
langchain
langchain-anthropic
langchain-groq
flask
ipykernel
python-dotenv
streamlit
-e .